new changes added for the graphs

+++ src/resolvers/restaurant/deleteRestaurant.js
@@ -1,14 +1,12 @@
-const { ApolloError } = require("apollo-server-express");
+const { ApolloError, AuthenticationError } = require("apollo-server-express");
 const { Restaurant } = require("../../models");
 
-const deleteRestaurant = async (_, { restaurantId }) => {
+const deleteRestaurant = async (_, { restaurantId }, { user }) => {
   try {
-    await Restaurant.findByIdAndDelete(restaurantId);
-    return {
-      success: true,
-      status: 0,
-      message: "Trip deleted successfully",
-    };
+    if (!user) {
+      throw new AuthenticationError("Unauthorised to perform this operation");
+    }
+    return await Restaurant.findByIdAndDelete(restaurantId);
   } catch (error) {
     console.log(`[ERROR]: Failed delete trip | ${error.message}`);
     throw new ApolloError("Failed delete trip");
